% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testo.R
\name{completemydata}
\alias{completemydata}
\title{Complete cases in mydata}
\usage{
completemydata()
}
\value{
Complete mydata
}
\description{
Complete cases in mydata
}
\examples{
\dontrun{
   completemydata()
}
}
