% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testo.R
\name{CrossVRateplan}
\alias{CrossVRateplan}
\title{Plot th CrossTable for RatePlane Variable with respect to Status}
\usage{
CrossVRateplan()
}
\value{
CrossTable
}
\description{
Plot th CrossTable for RatePlane Variable with respect to Status
}
\examples{
\dontrun{
   CrossVRateplan()
}
}
