% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_acq.R
\name{data_acq}
\alias{data_acq}
\title{Data acquisition}
\usage{
data_acq(id)
}
\arguments{
\item{id}{Symbol selected for data acquisition}
}
\value{
data.frame
}
\description{
Acquisition of data from 'oanda' website through quantmod package
}
\examples{
# Data acquisition currency rate between Euro and US dollars
data_acq("EUR/USD")

}
