% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testo.R
\name{sapplymydata}
\alias{sapplymydata}
\title{Missing values in mydata}
\usage{
sapplymydata()
}
\value{
Number of missing values for each column
}
\description{
Missing values in mydata
}
\examples{
\dontrun{
   sapplymydata()
}
}
